Inline KvStoreDatabase
